Refactor game types
